Refactor project setup by moving CLI project creation logic to a separate function and simplifying main.rs. This change enhances code organization and readability.
